requirements.txt
python-dotenv
pdfplumber
python-docx
pydantic
openai 
requests
gspread
oauth2client
pandas